updated readme.md, refactoring.

+++ controllers/ownerM.js
@@ -15,29 +15,51 @@ module.exports = {
   },
 
   newOwnerM: async (req, res, next) => {
-    const newOwnerM = new OwnerM(req.value.body);
-    const ownerM = await newOwnerM.save();
-    res.status(201).json(ownerM);
+    if(process.env.NODE_ENV && process.env.NODE_ENV == 'production') {
+      res.status(200).json('disabled');
+    }
+    else {
+      const newOwnerM = new OwnerM(req.value.body);
+      const ownerM = await newOwnerM.save();
+      res.status(201).json(ownerM);
+    }
   },
 
   deleteOwnerM: async (req, res, next) => {
-    const { ownermid } = req.value.params;
-    const ownerM = await OwnerM.findByIdAndRemove(ownermid);
-    res.status(200).json({ success: true });
+    if(process.env.NODE_ENV && process.env.NODE_ENV == 'production') {
+      res.status(200).json('disabled');
+    }
+    else {
+      const { ownermid } = req.value.params;
+      const ownerM = await OwnerM.findByIdAndRemove(ownermid);
+      res.status(200).json({ success: true });
+    }
   },
 
   replaceOwnerM: async (req, res, next) => {
-    const { ownermid } = req.value.params;
-    const newProps = req.value.body;
-    const ownerMToReplace = await OwnerM.findByIdAndUpdate(ownermid, newProps);
-    res.status(200).json({ success: true });
+    if(process.env.NODE_ENV && process.env.NODE_ENV == 'production') {
+      res.status(200).json('disabled');
+    }
+    else {
+      const { ownermid } = req.value.params;
+      const newProps = req.value.body;
+      const ownerMToReplace = await OwnerM.findByIdAndUpdate(ownermid, newProps);
+      const ownerMReplaced = await OwnerM.findById(ownermid);
+      res.status(200).json({ ownerMReplaced });
+    }
   },
 
   updateOwnerM: async (req, res, next) => {
-    const { ownermid } = req.value.params;
-    const newProps = req.value.body;
-    const ownerMToUpdate = await OwnerM.findByIdAndUpdate(ownermid, newProps);
-    res.status(200).json({ success: true });
+    if(process.env.NODE_ENV && process.env.NODE_ENV == 'production') {
+      res.status(200).json('disabled');
+    }
+    else {
+      const { ownermid } = req.value.params;
+      const newProps = req.value.body;
+      const ownerMToUpdate = await OwnerM.findByIdAndUpdate(ownermid, newProps);
+      const ownerMUpdated = await OwnerM.findById(ownermid);
+      res.status(200).json({ ownerMUpdated });
+    }
   }
 
 };
